requirements.txt
# requirements.txt
# CHF Prediction Project Dependencies

# Core Data Science Libraries
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
scipy>=1.10.0,<2.0.0

# Machine Learning Models
xgboost>=1.7.0,<2.0.0
lightgbm>=4.0.0,<5.0.0

# Deep Learning
torch>=2.0.0,<3.0.0
# Note: For GPU support, install torch with CUDA support:
# pip install torch --index-url https://download.pytorch.org/whl/cu118

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Data Download
kagglehub>=0.2.0

# Configuration and Logging
pyyaml>=6.0
colorlog>=6.7.0

# Progress Bars
tqdm>=4.65.0

# Optional: For enhanced features
# tensorboard>=2.13.0  # For TensorBoard logging
# plotly>=5.14.0       # For interactive plots
# jupyterlab>=4.0.0    # For notebook development

# Development Tools (optional)
# pytest>=7.3.0        # For testing
# pytest-cov>=4.1.0    # For coverage
# black>=23.0.0        # For code formatting
# flake8>=6.0.0        # For linting
# mypy>=1.3.0          # For type checking